Move promotions simulator to first level menu
